+++ src/pages/DocumentDisplay.vue
@@ -69,7 +69,7 @@
             anchor="bottom middle"
             self="top middle"
           >
-            Add new document with the current one as parent
+            Add a new document with the currently opened one as a parent
           </q-tooltip>
         </q-btn>